Makefile

+++ ft_memccpy.c
@@ -6,7 +6,7 @@
 /*   By: pmoyaha <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/30 10:21:05 by pmoyaha           #+#    #+#             */
-/*   Updated: 2019/05/30 11:29:50 by pmoyaha          ###   ########.fr       */
+/*   Updated: 2019/06/04 10:30:06 by event            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memccpy(void *dst, const void *src, int c, size_t n)
 {
-	int				i;
+	size_t				i;
 	unsigned char	*dest;
 	unsigned char	*source;
 	unsigned char	chr;
